Relocate Storage parsing logic to new StorageParser class to follow SRP
