+++ src/main/java/ducats/Storage.java
@@ -25,13 +25,11 @@ import java.util.Scanner;
  */
 public class Storage {
 
-    protected File file;
-    protected File dataFolder;
+    private File dataFolder;
     private String dirpath;
+    private StorageParser storageParser;
 
     private String fileDelimiter = System.getProperty("file.separator");
-    //private String currentDir = System.getProperty("user.dir");
-
 
     /**
      * Constructor for the duke.Storage class, used to store songs into persistent storage in the form of a .txt file.
@@ -41,6 +39,7 @@ public class Storage {
     public Storage(String dirpath) {
         this.dirpath = dirpath;
         dataFolder = new File(dirpath);
+        storageParser = new StorageParser();
     }
 
     /**
@@ -103,7 +102,6 @@ public class Storage {
                 }
             }
         } catch (Exception e) {
-            System.out.println("que?");
             throw new DucatsException("", "io");
         }
         return result;
@@ -118,8 +116,7 @@ public class Storage {
     private void loadSongToList(SongList songList, String filepath) throws DucatsException {
         // loads data into list
         ArrayList<String> data = readStringsFromFile(filepath);
-        songList.add(convertSongFromString(data));
-
+        songList.add(storageParser.convertSongFromString(data));
     }
 
     /**
@@ -132,132 +129,8 @@ public class Storage {
     public void loadToList(SongList songList) throws DucatsException {
         for (final File songFile: dataFolder.listFiles()) {
             String filepath = dirpath + fileDelimiter + songFile.getName();
-            System.out.println(filepath);
             loadSongToList(songList, filepath);
         }
-    }
-
-    /**
-     * A function that converts an ArrayList of Strings into a Song object.
-     *
-     * @param s the ArrayList of Strings to be converted to a song.
-     * @throws DucatsException in the case of IO exceptions
-     */
-    private Song convertSongFromString(ArrayList<String> s) throws DucatsException {
-        if (s.size() == 0) {
-            throw new DucatsException("io","");
-        }
-        String[] sections = s.get(0).split(" ");
-        String name = sections[0];
-        String key = sections[1];
-        int tempo = Integer.parseInt(sections[2]);
-        Song song = new Song(name, key, tempo);
-        int i = 1;
-        if (s.size() > 1) {
-            while (i < s.size() && !s.get(i).equals("groups:")) {
-                song.addBar(convertBarFromString(s.get(i), i));
-                i++;
-            }
-            for (int j = i + 1; j < s.size(); j++) {
-                song.addGroup(convertGroupFromString(s.get(j)));
-            }
-        }
-        return song;
-    }
-
-    private Group convertGroupFromString(String s) throws DucatsException {
-        String[] rawBars = s.split(" ");
-        String name = rawBars[0];
-        ArrayList<Bar> bars = new ArrayList<>();
-        for (int i = 1; i < rawBars.length; i++) {
-            bars.add(convertBarFromString(rawBars[i], -1));
-        }
-        Group group = new Group(name, bars);
-        return group;
-    }
-
-    private Bar convertBarFromString(String s, int barIndex) throws DucatsException {
-        String barData = s.substring(1, s.length() - 1);
-        String[] rawChords = barData.split(",");
-        ArrayList<Chord> chords = new ArrayList<>();
-        for (String rawChord: rawChords) {
-            chords.add(convertChordFromString(rawChord));
-        }
-        Bar bar = new Bar(barIndex, chords);
-        return bar;
-    }
-
-    private Chord convertChordFromString(String s) throws DucatsException {
-        String noteData = s.substring(1, s.length() - 1);
-        String[] rawNotes = noteData.split(";");
-        ArrayList<Note> notes = new ArrayList<>();
-        for (String rawNote: rawNotes) {
-            notes.add(convertNoteFromString(rawNote));
-        }
-        Chord chord = new Chord(notes);
-        return chord;
-    }
-
-    private Note convertNoteFromString(String s) throws DucatsException {
-        String duration = "8"; //duration for each chord
-        boolean isStart = false;
-        if (s.length() == 3) { //unit note is the start of the note
-            isStart = true;
-        }
-        Pitch pitch;
-        switch (s.length() == 3 ? s.substring(0,2) : s) {
-        case "LC":
-            pitch = Pitch.LOWER_C;
-            break;
-        case "LD":
-            pitch = Pitch.LOWER_D;
-            break;
-        case "LE":
-            pitch = Pitch.LOWER_E;
-            break;
-        case "LF":
-            pitch = Pitch.LOWER_F;
-            break;
-        case "LG":
-            pitch = Pitch.LOWER_G;
-            break;
-        case "LA":
-            pitch = Pitch.LOWER_A;
-            break;
-        case "LB":
-            pitch = Pitch.LOWER_B;
-            break;
-        case "MC":
-            pitch = Pitch.MIDDLE_C;
-            break;
-        case "UD":
-            pitch = Pitch.UPPER_D;
-            break;
-        case "UE":
-            pitch = Pitch.UPPER_E;
-            break;
-        case "UF":
-            pitch = Pitch.UPPER_F;
-            break;
-        case "UG":
-            pitch = Pitch.UPPER_G;
-            break;
-        case "UA":
-            pitch = Pitch.UPPER_A;
-            break;
-        case "UB":
-            pitch = Pitch.UPPER_B;
-            break;
-        case "UC":
-            pitch = Pitch.UPPER_C;
-            break;
-        case "RT":
-            pitch = Pitch.REST;
-            break;
-        default:
-            throw new DucatsException("data","");
-        }
-        return new Note(duration, pitch, isStart);
     }
 
     /**
@@ -269,9 +142,8 @@ public class Storage {
     public void updateFile(SongList songList) throws DucatsException {
         for (Song song: songList.getSongList()) {
             String filepath = dirpath + fileDelimiter + song.getName() + ".txt";
-            System.out.println(filepath);
             createFile(filepath);
-            writeStringsToFile(song.toStringArrayList(), filepath);
+            writeStringsToFile(storageParser.getArrayList(song), filepath);
         }
     }
